+++ include/visual_servo/parameterTeleop.h
@@ -15,13 +15,23 @@ private:
     ros::NodeHandle nh_;
     ros::Rate * loopRate_;
     std::vector<boost::thread> threads_;
+    std::vector<std::string> NUMBER_PARAMETER_NAMES;
+    std::vector<std::string> STRING_PARAMETER_NAMES;
+    std::vector<double> parameters_;
+    std::vector<std::string> stringParameters_;
 public:
-
+    const std::vector<double>& parameters() const {return parameters_;}
+    const std::vector<std::string>& stringParameters() const {return stringParameters_;};
     ParameterListener();
     virtual ~ParameterListener();
     void ParameterLoop(std::string parameterName,int index,bool isString);
     void registerParameterCallback(const std::vector<std::string> parameterNames,bool isString);
-    std::vector<double> parameters;
-    std::vector<std::string> stringParameters;
+    bool getParameterValueViaName(const std::string parameterName,double &value);
+    bool getParameterValueViaName(const std::string parameterName,std::string &value);
+    bool getParameterValueViaName(const std::string parameterName,int &value);
+    bool getParameterValueViaName(const std::string parameterName,bool &value);
+    double getNumberParameterValueViaIndex(int index);
+    std::string getStringParameterValueViaIndex(int index);
+
 
 };